refactor(oxlint,oxfmt): Use `tracing` logger for LSP (#17707)
Fixes #17702

NOTE: Environment variable has changed from `RUST_LOG` to `OXC_LOG`.
