+++ apps/oxfmt/Cargo.toml
@@ -41,7 +41,6 @@ cow-utils = { workspace = true }
 editorconfig-parser = { workspace = true }
 ignore = { workspace = true, features = ["simd-accel"] }
 json-strip-comments = { workspace = true }
-log = { workspace = true }
 miette = { workspace = true }
 phf = { workspace = true, features = ["macros"] }
 rayon = { workspace = true }
