+++ apps/oxfmt/src/cli/init.rs
@@ -44,23 +44,3 @@ pub fn init_rayon(threads: Option<usize>) {
 
     rayon::ThreadPoolBuilder::new().num_threads(thread_count).build_global().unwrap();
 }
-
-/// To debug `oxc_formatter`:
-/// `OXC_LOG=oxc_formatter oxfmt`
-/// # Panics
-pub fn init_tracing() {
-    use tracing_subscriber::{filter::Targets, prelude::*};
-
-    // Usage without the `regex` feature.
-    // <https://github.com/tokio-rs/tracing/issues/1436#issuecomment-918528013>
-    tracing_subscriber::registry()
-        .with(std::env::var("OXC_LOG").map_or_else(
-            |_| Targets::new(),
-            |env_var| {
-                use std::str::FromStr;
-                Targets::from_str(&env_var).unwrap()
-            },
-        ))
-        .with(tracing_subscriber::fmt::layer())
-        .init();
-}
