+++ apps/oxfmt/src/cli/mod.rs
@@ -6,9 +6,10 @@ mod result;
 mod service;
 mod walk;
 
+pub use crate::core::utils::init_tracing;
 #[cfg(feature = "napi")]
 pub use command::MigrateSource;
 pub use command::{FormatCommand, Mode, format_command};
 pub use format::FormatRunner;
-pub use init::{init_miette, init_rayon, init_tracing};
+pub use init::{init_miette, init_rayon};
 pub use result::CliRunResult;
