+++ apps/oxfmt/src/core/utils.rs
@@ -1,5 +1,29 @@
 use std::{fs, io, io::Write, path::Path};
 
+/// To debug `oxc_formatter`:
+/// `OXC_LOG=oxc_formatter oxfmt`
+/// # Panics
+pub fn init_tracing() {
+    use tracing_subscriber::{filter::Targets, prelude::*};
+
+    // Usage without the `regex` feature.
+    // <https://github.com/tokio-rs/tracing/issues/1436#issuecomment-918528013>
+    tracing_subscriber::registry()
+        .with(std::env::var("OXC_LOG").map_or_else(
+            |_| Targets::new(),
+            |env_var| {
+                use std::str::FromStr;
+                Targets::from_str(&env_var).unwrap()
+            },
+        ))
+        .with(
+            tracing_subscriber::fmt::layer()
+                // https://github.com/tokio-rs/tracing/issues/2492
+                .with_writer(std::io::stderr),
+        )
+        .init();
+}
+
 pub fn read_to_string(path: &Path) -> io::Result<String> {
     // `simdutf8` is faster than `std::str::from_utf8` which `fs::read_to_string` uses internally
     let bytes = fs::read(path)?;
