+++ apps/oxlint/Cargo.toml
@@ -42,8 +42,8 @@ bpaf = { workspace = true, features = ["autocomplete", "bright-color", "derive"]
 cow-utils = { workspace = true }
 ignore = { workspace = true, features = ["simd-accel"] }
 miette = { workspace = true }
-log = { workspace = true }
 napi = { workspace = true, features = ["async"], optional = true }
+tracing = { workspace = true }
 napi-derive = { workspace = true, optional = true }
 rayon = { workspace = true }
 rustc-hash = { workspace = true }
