+++ crates/oxc_language_server/src/lib.rs
@@ -22,8 +22,6 @@ pub async fn run_server(
     server_version: String,
     tools: Vec<Box<dyn ToolBuilder>>,
 ) {
-    env_logger::init();
-
     let stdin = tokio::io::stdin();
     let stdout = tokio::io::stdout();
